Create push_zeros_end.go
Problem: You need to push the zeros to the end of the array.
The easy solution is to create a new array, then add all the numbers except zeros and count the zeros, then make a second loop for the total number of zeros. However, this solution is not optimal because we already know the positions of the zeros, so we don't need to iterate again. Instead, we can use swapping, as the provided code.